+++ src/Vankosoft/CmsBundle/Model/QuickLink.php
@@ -1,8 +1,11 @@
 <?php namespace Vankosoft\CmsBundle\Model;
 
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Sylius\Component\Resource\Model\ToggleableTrait;
 use Vankosoft\CmsBundle\Model\Interfaces\QuickLinkInterface;
 use Vankosoft\ApplicationBundle\Model\Traits\TranslatableTrait;
+use Vankosoft\CmsBundle\Model\Interfaces\QuickLinksCategoryInterface;
 
 class QuickLink implements QuickLinkInterface
 {
@@ -18,9 +21,13 @@ class QuickLink implements QuickLinkInterface
     /** @var string */
     protected $linkPath;
     
+    /** @var Collection|QuickLinksCategory[] */
+    protected $categories;
+    
     public function __construct()
     {
         $this->fallbackLocale   = 'en_US';
+        $this->categories       = new ArrayCollection();
     }
     
     public function getId()
@@ -55,5 +62,30 @@ class QuickLink implements QuickLinkInterface
     public function isPublished(): ?bool
     {
         return $this->enabled;
+    }
+    
+    public function getCategories(): Collection
+    {
+        return $this->categories;
+    }
+    
+    public function addCategory( QuickLinksCategoryInterface $category ): self
+    {
+        if ( ! $this->categories->contains( $category ) ) {
+            $this->categories[] = $category;
+            $category->addQuickLink( $this );
+        }
+        
+        return $this;
+    }
+    
+    public function removeCategory( QuickLinksCategoryInterface $category ): self
+    {
+        if ( $this->categories->contains( $category ) ) {
+            $this->categories->removeElement( $category );
+            $category->removeQuickLink( $this );
+        }
+        
+        return $this;
     }
 }